Fix a number of bugs (#832)
* adds an expand and flatten command

* add petstore to canary tests (closes #665)

* fix invalid type name for nullable primitive in array

* add support full null type properties (closes #811)

* fix var name generation for vars with [] in them (closes #809)

* builtins and strfmts are not pointers in map values (closes #786)

* consider tags in client gen (closes #795)

* generate operations with more than 1 tag name in shared package (closes #500)

* support additionalProperties: true (closes #822)

* use more unique name for additional properties

* regen samples

+++ hack/run-canary.sh
@@ -8,13 +8,17 @@ fi
 
 for dir in $(ls fixtures/canary)
 do
-  pushd fixtures/canary/$dir
-  rm -rf client models restapi cmd
-  swagger generate client
-  go test ./...
-  if [ $dir != 'kubernetes' ] && [ $dir != 'ms-cog-sci' ] ; then
-    swagger generate server
+  if [ $dir != "bitbucket.org" ]; then
+    pushd fixtures/canary/$dir
+    rm -rf client models restapi cmd
+    swagger generate client
     go test ./...
+    if [ $dir != 'kubernetes' ] && [ $dir != 'ms-cog-sci' ] ; then
+      swagger generate server
+      go test ./...
+    fi
+    popd
+  else
+    echo "$dir is disabled for now"
   fi
-  popd
 done
